Move test of phone format checker to config package
