+++ pkg/util/validation/formats_test.go
@@ -11,17 +11,6 @@ func backgroundCtx() context.Context {
 	return context.Background()
 }
 
-func TestFormatPhone(t *testing.T) {
-	f := FormatPhone{}.CheckFormat
-
-	Convey("FormatPhone", t, func() {
-		So(f(backgroundCtx(), 1), ShouldBeNil)
-		So(f(backgroundCtx(), "+85298765432"), ShouldBeNil)
-		So(f(backgroundCtx(), ""), ShouldBeError, "not in E.164 format")
-		So(f(backgroundCtx(), "foobar"), ShouldBeError, "not in E.164 format")
-	})
-}
-
 func TestFormatEmail(t *testing.T) {
 	Convey("FormatEmail", t, func() {
 		f := FormatEmail{}.CheckFormat
